Add Shakespeare texts, fix src/test structure and fix a whitespace bug
Main source files and test files are now in the same package, so tests can test on protected members and methods. This was just a little Eclipse config, the actual file locations are the same. Also improved testing and fixed a bug where the document tokeniser was missing alternative whitespace because the split regex was " " and not "\\s".